+++ Makrowave-Type-Backend/Controllers/SettingsController.cs
@@ -1,29 +1,34 @@
-﻿using Makrowave_Type_Backend.Dtos;
+﻿using System.Security.Claims;
+using Makrowave_Type_Backend.Dtos;
 using Makrowave_Type_Backend.Models;
 using Makrowave_Type_Backend.Models.Entities;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 
 namespace Makrowave_Type_Backend.Controllers;
 
+[Route("api/[controller]")]
 [ApiController]
 public class SettingsController : ControllerBase
 {
     private readonly DatabaseContext _dbContext;
 
 
-    public SettingsController(DatabaseContext dbContext, DefaultTheme defaultTheme, IConfiguration config)
+    public SettingsController(DatabaseContext dbContext)
     {
         _dbContext = dbContext;
 
     }
     
-    
+    [Authorize(Policy = "SessionCookie", AuthenticationSchemes = "SessionCookie")]
     [HttpGet("theme")]
-    public async Task<ActionResult<string>> GetUserTheme(Guid userId)
+    public async Task<ActionResult<string>> GetUserTheme()
     {
+        var userId = await GetUserId(Request.Cookies["session"]);
+        Console.WriteLine(userId);
         if (!UserExists(userId))
         {
-            return NotFound();
+            return NotFound("User does not exist");
         }
         var theme = await _dbContext.UserThemes.FindAsync(userId);
         var gradientList = _dbContext.GradientColors.Where(color => color.UserThemeId == userId).OrderBy(color => color.Id).Select(color => color.Color).ToList();
@@ -41,19 +46,21 @@ public class SettingsController : ControllerBase
         };
         return Ok(result);
     }
-
+    [Authorize(Policy = "SessionCookie", AuthenticationSchemes = "SessionCookie")]
     [HttpPost("theme")]
-    public async Task<ActionResult<string>> SaveUserTheme(Guid userId, UserThemeDto userTheme)
+    public async Task<ActionResult<string>> SaveUserTheme(UserThemeDto userTheme)
     {
-        if (!UserExists(userId))
+        var userId = await GetUserId(Request.Cookies["session"]);
+        
+        if (!UserExists(userId) || !_dbContext.UserThemes.Any(theme => theme.UserThemeId == userId))
         {
-            return NotFound();
+            return NotFound("User does not exist");
         }
 
         var gradients = _dbContext.GradientColors.Where(color => color.UserThemeId == userId).ToList();
         _dbContext.GradientColors.RemoveRange(gradients);
         var theme = await _dbContext.UserThemes.FindAsync(userId);
-        theme.UiText = userTheme.UiText;
+        theme!.UiText = userTheme.UiText;
         theme.UiBackground = userTheme.UiBackground;
         theme.TextIncomplete = userTheme.TextIncomplete;
         theme.TextComplete = userTheme.TextComplete;
@@ -68,8 +75,18 @@ public class SettingsController : ControllerBase
         return Ok();
     }
     
-    bool UserExists(Guid userId)
+    private bool UserExists(Guid userId)
     {
         return _dbContext.Users.Any(u => u.UserId == userId);
+    }
+
+    private async Task<Guid> GetUserId(string sessionId)
+    {
+        Console.WriteLine(sessionId);
+        if (Guid.TryParse(sessionId, out var sessionGuid))
+        {
+            return (await _dbContext.Sessions.FindAsync(sessionGuid))!.UserId;
+        }
+        return Guid.Empty;
     }
 }
